+++ fold_test.ts
@@ -3,7 +3,7 @@ import {
   assertRejects,
 } from "https://deno.land/std@0.140.0/testing/asserts.ts";
 import { reduce } from "./fold.ts";
-import { getAsyncIterable } from "./testing_testing.ts";
+import { getAsyncIterable } from "./testing_test.ts";
 
 Deno.test("reduce()", async () => {
   assertEquals(